Pre alpha - openstack services are spawned in other namesapce

+++ scripts/namespace-cleaner.sh
@@ -2,25 +2,27 @@
 
 NAMESPACE=${NAMESPACE:-''}
 
+if [ -z "${NAMESPACE}" ]; then
+    echo "Please set NAMESPACE variable"
+    exit 1
+fi
+
 helm list -a | grep $NAMESPACE | awk '{print $1}' | xargs helm delete --purge
+#sleep 60
+
+
+TO_CLEAN="""all configmaps componentstatuses
+         customresourcedefinitions cronjobs deployments
+         daemonsets events endpoints events jobs.batch
+         limitranges pods podtemplates replicasets
+         secrets serviceaccounts services statefulsets"""
+
 kubectl get pv -n $NAMESPACE --no-headers | grep $NAMESPACE | awk '{print $1}' | xargs kubectl delete pv --wait=false
 kubectl get pvc -n $NAMESPACE --no-headers | grep $NAMESPACE | awk '{print $1}' | xargs kubectl delete pvc --wait=false
-kubectl get all -n $NAMESPACE --no-headers | awk '{print $1}' | xargs kubectl -n $NAMESPACE --wait=false delete
-kubectl delete configmaps --all --namespace=$NAMESPACE
-#kubectl delete componentstatuses  --all --namespace=$NAMESPACE
-#kubectl delete customresourcedefinitions --all --namespace=$NAMESPACE
-kubectl delete cronjobs --all --namespace=$NAMESPACE
-kubectl delete deployments --all --namespace=$NAMESPACE
-kubectl delete daemonsets --all --namespace=$NAMESPACE
-kubectl delete events --all --namespace=$NAMESPACE
-kubectl delete endpoints --all --namespace=$NAMESPACE
-kubectl delete events --all --namespace=$NAMESPACE
-kubectl delete jobs.batch  --all --namespace=$NAMESPACE
-kubectl delete limitranges --all --namespace=$NAMESPACE
-kubectl delete pods --all --namespace=$NAMESPACE
-kubectl delete podtemplates --all --namespace=$NAMESPACE
-kubectl delete replicasets --all --namespace=$NAMESPACE
-kubectl delete secrets --all --namespace=$NAMESPACE
-kubectl delete serviceaccounts --all --namespace=$NAMESPACE
-kubectl delete services --all --namespace=$NAMESPACE
-kubectl delete statefulsets --all --namespace=$NAMESPACE
+
+for CLEAN in $TO_CLEAN; do
+    echo "checking $CLEAN"
+    RESOURCE=$(kubectl get $CLEAN -n $NAMESPACE --no-headers)
+    echo -e "Found resources: \n $RESOURE"
+    echo $RESOURCE | awk '{print $1}' | xargs kubectl delete $CLEAN --wait=false
+done
